Changing const unsigned int size to size_t size.

+++ cs2430/sorting/printarrays.c
@@ -10,7 +10,7 @@
 
 #include "printarrays.h"
 
-void printarray(const int array[], const unsigned int numElements)
+void printarray(const int array[], const size_t numElements)
 {
   int index;
